renamend kuzzleDocument.js -> Document.js
